feat: add AI-powered features - application management, CV builder, job matching, notifications, subscription system
- BE: Add Application, CvBuilder, Notification, Subscription, AiScoring, JobMatching services and controllers
- BE: Add vector search, cover letter generation, candidate search with AI
- BE: Add GlobalExceptionHandler, AsyncConfig, admin stats endpoint
- BE: Update Auth, Job, Company, Candidate services with enhanced logic
- FE: Add admin, cv-builder, pricing, subscription, recruiter candidates pages
- FE: Add NotificationBell, auth-cookies, middleware, constants, services layer

+++ BE/src/main/java/com/example/be/entity/Job.java
@@ -72,10 +72,10 @@ public class Job {
     private String city;
 
     @Column(name = "salary_min")
-    private Integer salaryMin;
+    private Long salaryMin;
 
     @Column(name = "salary_max")
-    private Integer salaryMax;
+    private Long salaryMax;
 
     @Column(length = 10)
     private String currency = "VND";
